requirements-dev.txt
black
fake-bpy-module
pre-commit
pythonnet~=3.0.3
